Avoid pip 25.1 (#125)
Recently, our CI started failing with a mysterious error
```
urllib.error.HTTPError: HTTP Error 403: SSL is required
```

It is quite elusive, but empirically seems to be related to
self-referential extras in `pyproject.toml` (i.e. defining
`syntheseus[X]` as a dependency for `syntheseus[Y]`), which we use to
avoid listing the same dependency several times.

This is a fairly new feature of `pip`, and not a very well-documented
one. Some report it started working in `pip 21.2`, although changelogs
don't mention it explicitly. Ever since we started picking up `pip 25.1`
instead of `pip 25.0`, the build stopped working. I verified that this
can be fixed by either pinning to `pip 25.1` _or_ unrolling the
self-referential extras to duplicate all the dependencies (e.g.
explicitly writing down `syntheseus-root-aligned==0.1.0` as part of
`root-aligned`, `all-single-step` and `all`).

It seems slightly better to temporarily avoid the newest `pip` (and hope
the issue gets resolved downstream) instead of cluttering our
`pyproject.toml` with a (now legacy) duplication pattern; hence this PR
does the former in an attempt to get CI working again.