Add a utility to load a `SmilesListInventory` from file (#124)
By far the most common way to provide a purchasable molecule inventory
is `SmilesListInventory`, and the most common way of constructing that
is from a file containing a list of SMILES, one per line. So far, the
code to read the SMILES, strip endlines, and construct the inventory,
lived in the search CLI. However, calling search from arbitrary Python I
had to duplicate it. This PR avoids this (minor) duplication by making
this code a `classmethod` of `SmilesListInventory`.