Refactor search CLI to make it easier to access programmatically (#126)
Our search CLI has so far been structured in a way that makes it hard to
call parts of it programmatically from outside; for example, the code to
parse and reorganize search algorithm hyperparameters was only present
inside `run_from_config`. This PR refactors things to extract this logic
out, and also separates arguments configuring the search algorithm
itself from other configuration options. Finally, it also changes the
type of the `search_algorithm` argument from `str` to an appropriate
`Enum`, which allows for failing slightly earlier when an unsupported
algorithm has been selected.