Implement option to expand target even when purchasable (#130)
This PR implements an `expand_purchasable_target` option, which forces
the search algorithm to expand the target molecule, even if it is
already purchasable. This is similar to `exclude_target_from_stock` from
`aizynthfinder`. The implementation is similar to the existing
`expand_purchasable_mols`: purchasability metadata is still set
according to the full inventory, but if the option is set then expansion
happens regardless.