Pin `sentencepiece` to `0.2.0` in the RootAligned dependency (#133)
Recently we had issues with the `0.2.1` update to `sentencepiece` (which
RootAligned depends on), causing environments depending on
`syntheseus-root-aligned` to stop building in certain contexts (not in
GitHub CI for this repo though). Nevertheless, I updated
`syntheseus-root-aligned` to pin `sentencepiece==0.2.0` which is what
we've been using previously; this PR pulls that in into `syntheseus`.