Allow for installing with newest pip (#136)
When pip 25.1 was released, we started seeing installation issues
related to the use of self-referential extras in pyproject.toml; #125
thus pinned pip to <25.1. Since then pip 25.2 was released, and I
revisited the issue. Pinning to neither 25.1 nor 25.1.1 reproduces the
old error message exactly, although both cases still fail during
installation due to an inability to resolve the env.

This PR adds a lower pin above the bad 25.1.x version to
environment_full.yml, and unpins in environment.yml (which still allows
for python 3.8; we can drop support for it soon and then also have the
same lower pin on pip).