Run single-step models CI on GPU (#139)
Even though our single-step models CI uses a GPU-enabled environment
(defined in `environment_full.yml`), it still runs on a standard CPU
runner, not actually verifying the models run on GPU. Through Microsoft,
we managed to get access to a pool of GPU runners (with one Tesla T4 GPU
per runner). This PR extends the models CI to run both on CPU and GPU.
Additionally, for the GPU version, it also explicitly asserts that a GPU
is available. As a side-effect, we can now also run tests which
inherently require a GPU (currently, there is only one tiny test with
this property, in `utils/test_parallel.py`, and one can verify it is
skipped in the CPU CI but ran in GPU CI).