Silence various warnings from single-step models (#141)
Under newer dependency versions (either newer `torch` from #137, or
newer `rdkit`), some new warnings are getting produced when running the
single-step models. These so far seem harmless as quantitative model
performance is the same as it was before. This PR silences these new
warnings. As a bonus, it also changes the way Graph2Edits was silencing
`rdkit` messages from global to local (to avoid changing global state of
the library just because a certain single-step model was loaded).